Add ability to specify allowed SqlLocalDb versions
When new LocalDb versions are released DbTestMonkey would be default
always attempt to use the latest version. The problem is that the Dac
binaries that were used to deploy would not always support new versions.
The LocalDb versions that DbTestMonkey will consider are now
configurable in the test projects app.config file, an example is below.

<sqlServer
isLocalDbInstance="true"
localDbInstanceName="DbTestMonkey">
<allowedLocalDbVersions>
<version>11.0</version>
<version>12.0</version>
</allowedLocalDbVersions>
<databases>
.. database configuration here
</databases>

Resolves: #19
See also: #18