Fix NUnit working directory and DbTestMonkey project build order
Currently the DbTestMonkey solution is failing to build and run unit
tests
due to a dependent project being built after the fact. There is
also an
issue where NUnit >= 3.0 no longer sets the
Environment.CurrentDirectory
which causes DACPAC detection to fail.

The
working directory is now pre-set in the OneTimeSetUp method for NUnit.